+++ tests/backspace-clear-cells/change.js
@@ -1,7 +1,7 @@
-module.exports = function(plugin, change) {
-    const { state } = change;
-    const blockStart = state.document.getDescendant('anchor');
-    const blockEnd = state.document.getDescendant('focus');
+export default function(plugin, change) {
+    const { value } = change;
+    const blockStart = value.document.getDescendant('anchor');
+    const blockEnd = value.document.getDescendant('focus');
 
     const withCursor = change
         .collapseToStartOf(blockStart)
@@ -9,10 +9,10 @@ module.exports = function(plugin, change) {
 
     return plugin.onKeyDown(
         {
+            key: 'Backspace',
             preventDefault() {},
             stopPropagation() {}
         },
-        { key: 'backspace' },
         withCursor
     );
-};
+}