+++ tests/backspace-collapsed/change.js
@@ -1,18 +1,17 @@
-module.exports = function(plugin, change) {
-    const { state } = change;
-    const blockStart = state.document.getDescendant('anchor');
+export default function(plugin, change) {
+    const { value } = change;
+    const blockStart = value.document.getDescendant('anchor');
 
-    const withCursor = change
-        .collapseToStartOf(blockStart);
+    const withCursor = change.collapseToStartOf(blockStart);
 
     plugin.onKeyDown(
         {
+            key: 'Backspace',
             preventDefault() {},
             stopPropagation() {}
         },
-        { key: 'backspace' },
         withCursor
     );
 
     return change;
-};
+}