+++ tests/insert-column-preserve-data/change.js
@@ -1,0 +1,7 @@
+export default function(plugin, change) {
+    const { value } = change;
+    const cursorBlock = value.document.getDescendant('_cursor_');
+    change.moveToRangeOf(cursorBlock);
+
+    return plugin.changes.insertColumn(change, 1);
+}