+++ tests/insert-table/change.js
@@ -1,9 +1,7 @@
-module.exports = function(plugin, change) {
-    const { state } = change;
-    const cursorBlock = state.document.getDescendant('_cursor_');
-    change
-        .moveToRangeOf(cursorBlock)
-        .move(6); // Cursor here: Before|After
+export default function(plugin, change) {
+    const { value } = change;
+    const cursorBlock = value.document.getDescendant('_cursor_');
+    change.moveToRangeOf(cursorBlock).move(6); // Cursor here: Before|After
 
     return plugin.changes.insertTable(change);
-};
+}