+++ tests/on-shift-tab-first-cell/change.js
@@ -1,19 +1,20 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
-    const cursorBlock = change.state.document.getDescendant('_cursor_');
+export default function(plugin, change) {
+    const cursorBlock = change.value.document.getDescendant('_cursor_');
     change.moveToRangeOf(cursorBlock);
 
     plugin.onKeyDown(
         {
+            key: 'Tab',
+            shiftKey: true,
             preventDefault() {},
             stopPropagation() {}
         },
-        { key: 'tab', isShift: true },
         change
     );
 
-    const position = plugin.utils.getPosition(change.state);
+    const position = plugin.utils.getPosition(change.value);
 
     // First row (new one)
     expect(position.getRowIndex()).toEqual(0);
@@ -21,4 +22,4 @@ module.exports = function(plugin, change) {
     expect(position.getColumnIndex()).toEqual(2);
 
     return change;
-};
+}