+++ tests/remove-table-and-move-to-previous-block/change.js
@@ -1,0 +1,13 @@
+import expect from 'expect';
+
+export default function(plugin, change) {
+    const cursorBlock = change.value.document.getDescendant('_cursor_');
+    change.moveToRangeOf(cursorBlock);
+
+    plugin.changes.removeTable(change);
+    expect(change.value.startBlock.key).toEqual('_cursor_after_');
+    expect(change.value.startOffset).toEqual(
+        change.value.startBlock.text.length
+    );
+    return change;
+}