+++ tests/undo-insert-row/change.js
@@ -1,16 +1,16 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
-    const cursorBlock = change.state.document.getDescendant('_cursor_');
-    const initial = change.state.change({ save: false });
+export default function(plugin, change) {
+    const cursorBlock = change.value.document.getDescendant('_cursor_');
+    const initial = change.value.change({ save: false });
     initial.moveToRangeOf(cursorBlock);
-    const toTest = initial.state.change();
+    const toTest = initial.value.change();
     plugin.changes.insertRow(toTest);
 
     toTest.undo();
 
     // Back to previous cursor position
-    expect(toTest.state.startBlock.text).toEqual('Col 1, Row 1');
+    expect(toTest.value.startBlock.text).toEqual('Col 1, Row 1');
 
     return toTest;
-};
+}