+++ tests/undo-remove-column/change.js
@@ -1,11 +1,10 @@
-
-module.exports = function(plugin, change) {
-    const cursorBlock = change.state.document.getDescendant('_cursor_');
-    const initial = change.state.change({ save: false });
+export default function(plugin, change) {
+    const cursorBlock = change.value.document.getDescendant('_cursor_');
+    const initial = change.value.change({ save: false });
     initial.moveToRangeOf(cursorBlock);
-    const toTest = initial.state.change();
+    const toTest = initial.value.change();
     plugin.changes.removeColumn(toTest);
     toTest.undo();
 
     return toTest;
-};
+}